requirements.txt
# Function dependencies, for example:
# package>=version
numpy==1.21.5
pandas==1.4.2
google-cloud-bigquery==2.34.1S